Pull request #346: Fix and add x64 build to CI
Merge in ADGUARD-CORE-LIBS/vpn-libs from fix/x64_build to master

Squashed commit of the following:

commit 3f8f10c5c02c3e305448c4c427897c0c93044555
Author: Nikita Gorskikh <n.gorskikh@adguard.com>
Date:   Wed Sep 4 18:56:17 2024 +0300

    Reuse python venv

commit cc5ad878f08b126c4fdb275355d97c11419bdaa4
Author: Nikita Gorskikh <n.gorskikh@adguard.com>
Date:   Wed Sep 4 14:42:56 2024 +0300

    Fix calling a batch script from another batch script

commit b339ad4045db846745969911852fb5039832bec0
Author: Nikita Gorskikh <n.gorskikh@adguard.com>
Date:   Wed Sep 4 14:40:58 2024 +0300

    Fix ctest not being in path

commit ec262ac7aca2453e8a4a343d8b5fceeac93479dc
Author: Nikita Gorskikh <n.gorskikh@adguard.com>
Date:   Wed Sep 4 13:03:43 2024 +0300

    Build vpnlibs_core for x64 during CI

commit fde6c507dc1c75d4e130eebc9d3e884bf0580c04
Author: Nikita Gorskikh <n.gorskikh@adguard.com>
Date:   Wed Sep 4 13:03:21 2024 +0300

    Fix x64 build and some clang-tidy warnings

+++ core/src/tun_device_listener.cpp
@@ -2,8 +2,15 @@
 
 #include <algorithm>
 #include <cassert>
+#include <cstddef>
+#include <cstdint>
+#include <mutex>
 #include <numeric>
 #include <span>
+#include <utility>
+
+#include <vpn/event_loop.h>
+#include <vpn/vpn.h>
 
 #define log_conn(lstnr_, id_, lvl_, fmt_, ...)                                                                         \
     lvl_##log((lstnr_)->m_log, "[L:{}] " fmt_, (uint64_t) (id_), ##__VA_ARGS__)
@@ -427,9 +434,9 @@ void TunListener::recv_packets_task(void *arg, ag::TaskId) {
         listener->m_recv_packets_queue.emplace_back(std::move(*packet));
     }
 
-    static constexpr size_t BATCH_SIZE = 64;
-    size_t count = std::min(listener->m_recv_packets_queue.size(), BATCH_SIZE);
-    listener->vpn->process_client_packets({.data = listener->m_recv_packets_queue.data(), .size = count});
+    static constexpr int BATCH_SIZE = 64;
+    int count = (int) std::min((size_t) listener->m_recv_packets_queue.size(), (size_t) BATCH_SIZE);
+    listener->vpn->process_client_packets({.data = listener->m_recv_packets_queue.data(), .size = (unsigned) count});
     listener->m_recv_packets_queue.erase(
             listener->m_recv_packets_queue.begin(), listener->m_recv_packets_queue.begin() + count);
